The filtering approach requires enumerating all available environment variables. This makes all these variables inputs to the configuration cache and produces a ton of warnings when environment variable reads are tracked, effectively breaking the build and preventing tracking code from landing in gradle/gradle.
This commit does two things. First, the enumeration of the variables and
name-based filtering is replaced with the explicit list of variables
because of the limitations of the provider API. Currently, it isn't
possible to ask for a subset of environment variables. Unfortunately,
this makes adding new environment variables a bit harder. However, the
new API to request a subset of variables based on the prefix is
currently considered.

Second, the creation of the environment variable providers is optimized.
Currently, 69 variables are read in ~8000 test tasks. The naive approach
results in 560K providers being created; this blows up memory usage and
configuration cache disk footprint to the point of OOM. A build service
is introduced to reuse filtered variable values for all test tasks, so
only the first read requires creating a provider.